pyproject.toml
[project]
name = "YokaiEval"
version = "0.1.0"
description = "Yokai Knowledge Evaluation Dataset"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.2.0",
    "beautifulsoup4>=4.12.3",
    "editables>=0.5",
    "hatchling>=1.27.0",
    "openai>=1.55.0",
    "pandas>=2.2.3",
    "peft>=0.14.0",
    "protobuf>=5.29.1",
    "pylcs>=0.1.1",
    "python-dotenv>=1.0.1",
    "rapidfuzz>=3.12.1",
    "requests>=2.32.3",
    "sentencepiece>=0.2.0",
    "setuptools>=75.6.0",
    "torch>=2.5.1",
    "transformers>=4.46.3",
    "wheel>=0.45.1",
]

[tool.ruff]
line-length = 88
